README Class Fraction
This code was taught to me taken from a project in school inspired by the book Absolute C++ 5th Edition. Fractions that can be reduced using GCD. A class is a description of how objects behaves. An object is an instance of the class that follows the description. This design class Fraction is intended to learn and practice how to code from a fraction to a fraction is inputted by user as it is intended a/b and the program will result in the same manner.